requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp>=3.8.1
gspread>=5.7.0
google-auth>=2.20.0
google-auth-oauthlib>=0.8.0
google-api-python-client>=2.70.0
psutil>=5.9.5
tzdata>=2025.3